Added license automatically. Controller code, API structure.
