+++ config/crd/bases/pathways-api.pathways.domain_pathwaysapis.yaml
@@ -1,3 +1,17 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 ---
 apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
@@ -38,6 +52,17 @@ spec:
             type: object
           spec:
             properties:
+              colocationPolicy:
+                description: |-
+                  ColocationPolicy defines whether the user job and the Pathways resources (RM, proxy)
+                  must be colocated on TPUs with the Pathways workers or not.
+                  Users may opt for best-effort placement where scheduler places the RM and proxy
+                  on the CPU nodepools by default.
+                  Default is best-effort.
+                enum:
+                - colocate
+                - best-effort
+                type: string
               maxRestarts:
                 description: Maximum number of times the JobSet is restarted.
                 format: int32
@@ -46,10 +71,6 @@ spec:
                 description: Number of TPU slices requested for the Pathways workers.
                 format: int32
                 type: integer
-              pathwaysClientVersion:
-                description: PathwaysClientVersion is the version of the Pathways
-                  client.
-                type: string
               pathwaysControllerNodeSelector:
                 additionalProperties:
                   type: string
@@ -59,8 +80,12 @@ spec:
                 type: object
               pathwaysDir:
                 description: |-
-                  PathwaysDir is the GCS location at which Pathways artifacts
-                  can be stored.
+                  PathwaysDir is a persistent location like GCS at which temporary
+                  Pathways artifacts can be stored like HBM state during interruptions.
+                  Currently, Pathways supports a precreated GCS directory only.
+                type: string
+              pathwaysVersion:
+                description: PathwaysVersion is the version of the Pathways client.
                 type: string
               pathwaysWorkerNodeSelector:
                 additionalProperties:
@@ -7998,6 +8023,10 @@ spec:
             description: PathwaysAPIStatus defines the observed state of PathwaysAPI
             properties:
               conditions:
+                description: |-
+                  Track the state of the Pathways workload, acceptable values are -
+                  Running, Suspended, Completed, Failed.
+                  Contains a human readable message to provide additional details to the // user.
                 items:
                   description: "Condition contains details for one aspect of the current
                     state of this API Resource.\n---\nThis struct is intended for
@@ -8066,11 +8095,6 @@ spec:
                   - type
                   type: object
                 type: array
-              workloadState:
-                description: |-
-                  Track the state of the Pathways workload, acceptable values are -
-                  Suspended, Completed, Failed
-                type: string
             type: object
         type: object
     served: true
