+++ config/samples/pathways-api_v1_pathwaysapi.yaml
@@ -1,12 +1,23 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 apiVersion: pathways-api.pathways.domain/v1
 kind: PathwaysAPI
 metadata:
-  labels:
-    app.kubernetes.io/name: pathways-api
-    app.kubernetes.io/managed-by: kustomize
   name: pathwaysapi-sample
 spec:
-  workloadName: "roshani-in-20"
+  workloadName: "roshani-in-27"
   pathwaysWorkerNodeSelector:
     cloud.google.com/gke-tpu-accelerator: tpu-v4-podslice
     cloud.google.com/gke-tpu-topology: 2x2x2
